move cahce to utils.cache and add PlayerProfileCache
