+++ src/main/java/fr/openmc/core/features/city/menu/main/buttons/ManageButton.java
@@ -7,7 +7,7 @@ import fr.openmc.core.features.city.CityManager;
 import fr.openmc.core.features.city.CityPermission;
 import fr.openmc.core.features.city.menu.CityModifyMenu;
 import fr.openmc.core.features.city.sub.milestone.rewards.MemberLimitRewards;
-import fr.openmc.core.utils.CacheOfflinePlayer;
+import fr.openmc.core.utils.cache.CacheOfflinePlayer;
 import fr.openmc.core.utils.messages.MessageType;
 import fr.openmc.core.utils.messages.MessagesManager;
 import fr.openmc.core.utils.messages.Prefix;
