+++ src/main/java/fr/openmc/core/features/economy/menu/TransactionsMenu.java
@@ -7,7 +7,7 @@ import fr.openmc.api.menulib.utils.StaticSlots;
 import fr.openmc.core.features.economy.Transaction;
 import fr.openmc.core.features.economy.TransactionsManager;
 import fr.openmc.core.items.CustomItemRegistry;
-import fr.openmc.core.utils.CacheOfflinePlayer;
+import fr.openmc.core.utils.cache.CacheOfflinePlayer;
 import net.kyori.adventure.text.Component;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
