+++ src/main/java/fr/openmc/core/features/friend/FriendRequest.java
@@ -1,7 +1,7 @@
 package fr.openmc.core.features.friend;
 
 import fr.openmc.core.OMCPlugin;
-import fr.openmc.core.utils.CacheOfflinePlayer;
+import fr.openmc.core.utils.cache.CacheOfflinePlayer;
 import fr.openmc.core.utils.messages.MessageType;
 import fr.openmc.core.utils.messages.MessagesManager;
 import fr.openmc.core.utils.messages.Prefix;
