+++ src/main/java/fr/openmc/core/utils/ItemUtils.java
@@ -1,10 +1,9 @@
 package fr.openmc.core.utils;
 
 import dev.lone.itemsadder.api.CustomStack;
-import fr.openmc.api.menulib.Menu;
-import fr.openmc.api.menulib.utils.ItemBuilder;
 import fr.openmc.core.features.mailboxes.MailboxManager;
 import fr.openmc.core.items.CustomItemRegistry;
+import fr.openmc.core.utils.cache.CacheOfflinePlayer;
 import fr.openmc.core.utils.messages.MessageType;
 import fr.openmc.core.utils.messages.MessagesManager;
 import fr.openmc.core.utils.messages.Prefix;
